+++ src/tag/li.js
@@ -1,6 +1,7 @@
 const Tag = require('./tag').Tag;
 const ATag = require('./a').ATag;
 const LoginTag = require('./login').LoginTag;
+const LogoutTag = require('./logout').LogoutTag;
 
 /**
  * @typedef LiTag
@@ -28,9 +29,12 @@ function LiTagAttrs(value, textSearch) {
 function LiTag(children, attrs) {
     if (
         children.length !== 1 ||
-        !(children[0] instanceof ATag || children[0] instanceof LoginTag || typeof children[0] === 'string')
+        !(children[0] instanceof ATag || 
+        children[0] instanceof LoginTag ||
+        children[0] instanceof LogoutTag || 
+        typeof children[0] === 'string')
     ) {
-        throw Error('<li> must have 1 child (<a>, <login> or text)')
+        throw Error('<li> must have 1 child (<a>, <login>, <logout> or text)')
     }
     this.children = children;
     this.attrs = attrs;