+++ test/test_schema.js
@@ -38,11 +38,15 @@ describe('Test schema', function () {
                     '       <li>' +
                     '           <login on-success="/successPath1" on-failure="/failurePath1" />' +
                     '       </li>' +
+                    '       <li>' +
+                    '           <logout on-success="/logoutSuccessPath1" on-failure="/logoutFailurePath1" />' +
+                    '       </li>' +
                     '       <li text-search="route 3 route3">' +
                     '           <a href="/route3">Route 3</a>' +
                     '       </li>' +
                     '   </ul>' +
                     '   <login on-success="/successPath" on-failure="/failurePath" />' +
+                    '   <logout on-success="/logoutSuccessPath2" on-failure="/logoutFailurePath2" />' +
                     '</section>';
                 const rootTag = parser.loadHtml(undefined, html);
                 const response = Response.fromTag(rootTag);
@@ -59,6 +63,8 @@ describe('Test schema', function () {
                                 "method": null,
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": null,
                                 "src": "/videoPath",
                                 "alt": "alt video name"
@@ -70,6 +76,8 @@ describe('Test schema', function () {
                                 "method": "POST",
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": "/route1",
                                 "src": null,
                                 "alt": null
@@ -81,6 +89,8 @@ describe('Test schema', function () {
                                 "method": 'GET',
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": "/route2",
                                 "src": null,
                                 "alt": null
@@ -92,6 +102,8 @@ describe('Test schema', function () {
                                 "method": null,
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": null,
                                 "src": null,
                                 "alt": null
@@ -103,6 +115,8 @@ describe('Test schema', function () {
                                 "method": "GET",
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": "/form/new/",
                                 "src": "https://placekitten.com/600/600",
                                 "alt": "alt text 2"
@@ -114,6 +128,21 @@ describe('Test schema', function () {
                                 "method": null,
                                 "on_login_failure": "/failurePath1",
                                 "on_login_success": "/successPath1",
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
+                                "path": null,
+                                "src": null,
+                                "alt": null
+                            },
+                            {
+                                "type": "logout",
+                                "description": null,
+                                "text_search": null,
+                                "method": null,
+                                "on_login_failure": null,
+                                "on_login_success": null,
+                                "on_logout_failure":"/logoutFailurePath1",
+                                "on_logout_success":"/logoutSuccessPath1",
                                 "path": null,
                                 "src": null,
                                 "alt": null
@@ -125,6 +154,8 @@ describe('Test schema', function () {
                                 "method": 'GET',
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": "/route3",
                                 "src": null,
                                 "alt": null
@@ -136,6 +167,21 @@ describe('Test schema', function () {
                                 "method": null,
                                 "on_login_failure": '/failurePath',
                                 "on_login_success": '/successPath',
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
+                                "path": null,
+                                "src": null,
+                                "alt": null
+                            },
+                            {
+                                "type": "logout",
+                                "description": null,
+                                "text_search": null,
+                                "method": null,
+                                "on_login_failure": null,
+                                "on_login_success": null,
+                                "on_logout_failure":'/logoutFailurePath2',
+                                "on_logout_success":'/logoutSuccessPath2',
                                 "path": null,
                                 "src": null,
                                 "alt": null
@@ -239,6 +285,8 @@ describe('Test schema', function () {
                                 "method": null,
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": null,
                                 "src": null,
                                 "alt": null                                
@@ -250,6 +298,8 @@ describe('Test schema', function () {
                                 "method": null,
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": null,
                                 "src": null,
                                 "alt": null
@@ -261,6 +311,8 @@ describe('Test schema', function () {
                                 "method": null,
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": null,
                                 "src": null,
                                 "alt": null
@@ -272,6 +324,8 @@ describe('Test schema', function () {
                                 "method": null,
                                 "on_login_failure":null,
                                 "on_login_success":null,
+                                "on_logout_failure":null,
+                                "on_logout_success":null,
                                 "path": null,
                                 "src": null,
                                 "alt": null
@@ -334,6 +388,8 @@ describe('Test schema', function () {
                                 "max_value_error": "Too high",
                                 "on_login_success":null,
                                 "on_login_failure":null,
+                                "on_logout_success":null,
+                                "on_logout_failure":null,
                                 "step": null,
                                 "meta": {
                                     "auto_select": false,
@@ -414,6 +470,8 @@ describe('Test schema', function () {
                                 "max_value_error": null,
                                 "on_login_success":null,
                                 "on_login_failure":null,
+                                "on_logout_success":null,
+                                "on_logout_failure":null,
                                 "step": null,
                                 "meta": {
                                     "auto_select": true,
@@ -451,6 +509,47 @@ describe('Test schema', function () {
                                 "max_value_error": null,
                                 "on_login_success": '/successPath',
                                 "on_login_failure": '/failurePath',
+                                "on_logout_success":null,
+                                "on_logout_failure":null,
+                                "step": null,
+                                "meta": {
+                                    "auto_select": false,
+                                    "multi_select": false,
+                                    "numbered": false
+                                },
+                                "method": null,
+                                "required": false,
+                                "default":null,
+                                "pattern": null,
+                                "status_exclude": false,
+                                "status_prepend": false,
+                                "url": null,
+                                "validate_type_error": null,
+                                "validate_type_error_footer": null,
+                                "validate_url": null
+                            },
+                            {
+                                "type": "logout",
+                                "name": "logout",
+                                "description": null,
+                                "header": null,
+                                "footer": null,
+                                "body": null,
+                                "value": null,
+                                "chunking_footer": null,
+                                "confirmation_label": null,
+                                "min_length": null,
+                                "min_length_error": null,
+                                "max_length": null,
+                                "max_length_error": null,
+                                "min_value": null,
+                                "min_value_error": null,
+                                "max_value": null,
+                                "max_value_error": null,
+                                "on_login_success": null,
+                                "on_login_failure": null,
+                                "on_logout_success": '/logoutSuccessPath',
+                                "on_logout_failure":'/logoutFailurePath',
                                 "step": null,
                                 "meta": {
                                     "auto_select": false,
@@ -528,6 +627,8 @@ describe('Test schema', function () {
                     "max_value_error": null,
                     "on_login_success":null,
                     "on_login_failure":null,
+                    "on_logout_success":null,
+                    "on_logout_failure":null,
                     "step": null,
                     "meta": {
                         "auto_select": false,
@@ -628,6 +729,8 @@ describe('Test schema', function () {
                     "max_value_error": null,
                     "on_login_success":null,
                     "on_login_failure":null,
+                    "on_logout_success":null,
+                    "on_logout_failure":null,
                     "step": null,
                     "meta": {
                         "auto_select": true,
@@ -656,7 +759,7 @@ describe('Test schema', function () {
                     return new FormItem({type: 'blabla'});
                 }
 
-                assert.throws(iThrow, Error, 'FormItem type="blabla" is not supported. Supported types: date,datetime,email,form-menu,float,hidden,int,location,login,range,regex,string,tel,phone,url,textarea');
+                assert.throws(iThrow, Error, 'FormItem type="blabla" is not supported. Supported types: date,datetime,email,form-menu,float,hidden,int,location,login,logout,range,regex,string,tel,phone,url,textarea');
             })
         })
     });
@@ -722,6 +825,8 @@ describe('Test schema', function () {
                             "max_value_error": null,
                             "on_login_success":null,
                             "on_login_failure":null,
+                            "on_logout_success":null,
+                            "on_logout_failure":null,
                             "step": null,
                             "meta": {
                                 "auto_select": true,
@@ -759,6 +864,8 @@ describe('Test schema', function () {
                             "max_value_error": null,
                             "on_login_success":null,
                             "on_login_failure":null,
+                            "on_logout_success":null,
+                            "on_logout_failure":null,
                             "step": 1,
                             "meta": {
                                 "auto_select": false,